Remove unecessary changes refactored to other PRs
