+++ samples/Sentry.Samples.Maui/MainPage.xaml.cs
@@ -32,19 +32,6 @@ public partial class MainPage
     {
         _count++;
 
-        var span = SentrySdk.StartSpan("button.click", "ui.action");
-        try
-        {
-            SentrySdk.ConfigureScope(scope => scope.Transaction ??= span.GetTransaction());
-            // Make an HTTP request to demonstrate Sentry's automatic HTTP tracking
-            using var httpClient = new HttpClient(new SentryHttpMessageHandler());
-            httpClient.GetAsync("https://www.example.com/").GetAwaiter().GetResult();
-            span.Status = SpanStatus.Ok;
-        }
-        finally
-        {
-            span.Finish();
-        }
         if (_count == 1)
         {
             CounterBtn.Text = $"Clicked {_count} time";
