+++ src/Sentry.Bindings.Android/Transforms/Metadata.xml
@@ -163,4 +163,7 @@
 
   <!-- Remove UpdateStatus to avoid CS0108 for conflicting _members fields in nested subclasses -->
   <remove-node path="/api/package[@name='io.sentry']/class[@name='UpdateStatus']" />
+
+  <!-- error CS0050: Inconsistent accessibility: return type 'INetworkBody' is less accessible than method 'NetworkBody.FromXxx' -->
+  <remove-node path="/api/package[@name='io.sentry.util.network']/interface[@name='NetworkBody']" />
 </metadata>
